moves standalone monitoring into the process package
