+++ pkg/process/monitor_external_test.go
@@ -1,6 +1,6 @@
 //go:build linux
 
-package monitor_test
+package process_test
 
 import (
 	"os/exec"
@@ -9,7 +9,7 @@ import (
 
 func TestExternalConsumer(t *testing.T) {
 	cmd := exec.CommandContext(t.Context(), "go", "test", "-mod=readonly", "./...")
-	cmd.Dir = "../../../examples/process-monitor"
+	cmd.Dir = "../../examples/process-monitor"
 	if output, err := cmd.CombinedOutput(); err != nil {
 		t.Fatalf("external consumer: %v\n%s", err, output)
 	}
